scripts/rtl_ingest/requirements.txt
# RTL ingest: 7-channel RTL-SDR -> Parakeet ASR -> feed
pyrtlsdr>=0.2.93
numpy>=1.20
scipy>=1.7
nemo_toolkit[asr]>=1.20.0
